Voice-ChatBot-Multilingual/requirement.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0   
anyio==4.2.0
attrs==23.2.0
certifi==2024.2.2        
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
distro==1.9.0
dnspython==2.5.0
email-validator==2.1.0.post1
fastapi==0.109.2
filelock==3.13.1
frozenlist==1.4.1
fsspec==2024.2.0
gTTS==2.5.1
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
llvmlite==0.42.0
MarkupSafe==2.1.5
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
numba==0.59.0
numpy==1.26.3
openai==0.28.0
openai-whisper==20231117
orjson==3.9.13
pandas==2.2.0
platformdirs==4.1.0
pyaes==1.6.1
pyasn1==0.5.1
pydantic==2.6.1
pydantic-extra-types==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.16.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2023.4
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rsa==4.9
setuptools==69.0.3
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
sympy==1.12
Telethon==1.33.1
tiktoken==0.6.0
torch==2.2.0
PS C:\Users\Ian Marx\Desktop\AI\Voice-ChatBot-Multilingual> python -m pip freeze requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0   
anyio==4.2.0
attrs==23.2.0
certifi==2024.2.2        
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
distro==1.9.0
dnspython==2.5.0
email-validator==2.1.0.post1
fastapi==0.109.2
filelock==3.13.1
frozenlist==1.4.1
fsspec==2024.2.0
gTTS==2.5.1
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
llvmlite==0.42.0
MarkupSafe==2.1.5
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
numba==0.59.0
numpy==1.26.3
openai==0.28.0
openai-whisper==20231117
orjson==3.9.13
pandas==2.2.0
platformdirs==4.1.0
pyaes==1.6.1
pyasn1==0.5.1
pydantic==2.6.1
pydantic-extra-types==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.16.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2023.4
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rsa==4.9
setuptools==69.0.3
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
sympy==1.12
Telethon==1.33.1
tiktoken==0.6.0
torch==2.2.0
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2023.4
ujson==5.9.0
urllib3==2.2.0
uvicorn==0.27.1
virtualenv==20.25.0
virtualenvwrapper-win==1.2.7
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4